Refactored GPT model
