+++ pyproject.toml
@@ -11,6 +11,8 @@ torch = "^2.0.0"
 tiktoken = "^0.3.2"
 numpy = "^1.24.2"
 pillow = "^9.4.0"
+tqdm = "^4.65.0"
+matplotlib = "^3.7.1"
 
 
 [tool.poetry.group.dev.dependencies]
