+++ src/models/base.py
@@ -1,0 +1,6 @@
+import torch.nn as nn
+
+
+class Model(nn.Module):
+    def __init__(self):
+        super().__init__()
