+++ src/modules/feed_forward.py
@@ -1,0 +1,15 @@
+import torch.nn as nn
+
+
+class FeedForward(nn.Module):
+    def __init__(self, n_embd: int, dropout: float = 0.2):
+        super().__init__()
+        self.net = nn.Sequential(
+            nn.Linear(n_embd, 4 * n_embd),
+            nn.ReLU(),
+            nn.Linear(4 * n_embd, n_embd),
+            nn.Dropout(dropout),
+        )
+
+    def forward(self, x):
+        return self.net(x)
